Update Left-Rotation.c
This file takes in as stdin two lines of space separated integers as follows
5 2
1 2 3 4 5
the first line are the amount of integers in the second line. The second number is the number of shifts left of the second line integers. 
The output of the function goes to stdout and it should be the shifted array of integers. The output for the input above would be:
4 5 1 2 3

cheers!