Create compare-the-triplets.c
This code takes in as input two lines of three space separated integers.
The task of the code is to assign a set (i.e. each line) of integers to Alice and one to Bob.
Each one of the integers is to be compared, and whom ever has the highest number at each index (i.e. 0,1, or 2 since they are three integer arrays) gets a "point". These points are tallied up, and Alice and Bob's total score is printed as a set of two space separated integers. Cheers.